Creo fichero para los enums, closes #48
Los enums que estaban en usuario y habilidad se pasan a un fichero aparte
