+++ lib/habilidad.ts
@@ -1,17 +1,4 @@
-export enum TipoHabilidad {
-    Posicionamiento,
-    Rotaciones,
-    Apoyo,
-  }
-
-export enum Nivel {
-    Muy_malo,
-    Malo,
-    Medio,
-    Bueno,
-    Muy_bueno,
-    Excelente,
-}
+import { TipoHabilidad, Nivel } from "./types";
 export class Habilidad {
     public tipo: TipoHabilidad;
     public nivel: Nivel;
