+++ lib/types.ts
@@ -1,0 +1,35 @@
+export enum Rol {
+    Fragger,
+    IGL,
+    Support,
+    Lurker,
+    Awper,
+  }
+
+export enum Actitud {
+    Agresivo,
+    Controlador,
+  }
+
+export enum Rango {
+    Plata,
+    Oro,
+    Ak,
+    Maestro,
+    EliteGlobal,
+  }
+
+export enum TipoHabilidad {
+    Posicionamiento,
+    Rotaciones,
+    Apoyo,
+  }
+
+export enum Nivel {
+    Muy_malo,
+    Malo,
+    Medio,
+    Bueno,
+    Muy_bueno,
+    Excelente,
+}
